[クラス削除、クラス名変更]既存のMagicクラスを削除>SpellListクラスをMagicと命名
