+++ Main/src/Game.java
@@ -1,6 +1,5 @@
 import character.Character;
 import character.hitPoint.HitPoint;
-import character.magic.Magic;
 import character.magic.SpellList;
 import character.attackPoint.AttackPoint;
 import character.name.Name;
